requirements.txt
django==5.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.10
django-cors-headers==4.7.0
python-dotenv==1.1.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
requests==2.32.4
requests-oauthlib==2.0.0
oauthlib==3.2.2